Removed useless spaces + renamed GetPahts.js to GetPaths.js (typo)
Should still build and work the same, not tested though
Please test before merging!
