+++ src/Checks/CheckProgressCurrentPage.js
@@ -1,7 +1,3 @@
-
-
-
-//Progress
 const {MapDrops} = require("../functions/MapDrops");
 const {GetRustDrops} = require("../functions/GetRustDrops");
 const data = require("../Data/SavedData");
@@ -21,8 +17,6 @@ async function CheckProgressCurrentPage(page, dropspage, startch, campaignpage) 
         if (data.claimed.includes(e.twitch_name)) {
             e.claimed = true;
         }
-
-
     })
 
     for (let i = 0; i < data.dropsmap.length; i++)
@@ -32,15 +26,9 @@ async function CheckProgressCurrentPage(page, dropspage, startch, campaignpage) 
         } else if (data.claimed.includes(name)) {
             return -1
         }
-
-
     }
-
-
-
 }
 
 module.exports = {
     CheckProgressCurrentPage
 }
-
